+++ GameEngine/GameEngine/Engine/Physic/DynamicMotionState.hpp
@@ -5,6 +5,7 @@
 #include <Utils/MatrixConversion.hpp>
 
 #include <glm/glm.hpp>
+#include <glm/gtc/type_ptr.hpp>
 #include <glm/gtc/matrix_transform.hpp>
 #include <Core/Link.hpp>
