+++ GameEngine/GameEngine/Engine/Systems/CameraSystem.hh
@@ -8,27 +8,15 @@
 #include <Core/Drawable.hh>
 #include <Core/RenderThread.hpp>
 
-namespace gl
-{
-	class Shader;
-	struct Uniform;
-	class UniformBlock;
-	class RenderPass;
-	class RenderOnScreen;
-	class Pipeline;
-}
-
-struct RenderKey
-{
-	gl::Key<gl::Shader> shader;
-	gl::Key<gl::UniformBlock> global_state;
-	gl::Key<gl::Uniform> pro_matrix;
-	gl::Key<gl::Uniform> view_matrix;
-	gl::Key<gl::Uniform> normal_matrix;
-	gl::Key<gl::RenderPass> renderPass;
-	gl::Key<gl::RenderOnScreen> renderOnScreen;
-	gl::Key<gl::Pipeline> pipeline;
-};
+//namespace gl
+//{
+//	class Shader;
+//	struct Uniform;
+//	class UniformBlock;
+//	class RenderPass;
+//	class RenderOnScreen;
+//	class Pipeline;
+//}
 
 class CameraSystem : public System
 {
@@ -36,7 +24,6 @@ public:
 	CameraSystem(std::weak_ptr<AScene> &&scene);
 	virtual ~CameraSystem(){}
 	
-	void setManager();
 	void setRenderDebugMode(bool t);
 	bool getRenderDebugMode() const;
 	void getRayFromMousePosOnScreen(glm::vec3 &from, glm::vec3 &to);
@@ -49,12 +36,7 @@ protected:
 	double	_totalTime;
 	EntityFilter _drawable;
 	EntityFilter _camera;
-	AGE::RenderThread *_renderThread;
-	gl::RenderManager *_renderManager;
 	
-	RenderKey key;
-
-	AGE::Vector<AGE::DrawableCollection> _drawList;
 	virtual void updateBegin(double time);
 	virtual void updateEnd(double time);
 	virtual void mainUpdate(double time);
